+++ lib/src/treeview.dart
@@ -8,7 +8,7 @@ class _TreeViewProps {
     required this.itemBuilder,
     required this.indicator,
     this.indicatorBuilder,
-    required this.rowExtent,
+    required this.childExtent,
     required this.spacing,
     required this.animationDuration,
   });
@@ -22,12 +22,24 @@ class _TreeViewProps {
   )?
   indicatorBuilder;
   final Widget indicator;
-  final double rowExtent;
+  final double childExtent;
   final double spacing;
   final Duration animationDuration;
 }
 
-enum Placement { above, below }
+///Indicates the placement of a node with respect to another node.
+///
+///Used in the [TreeView.indicatorBuilder] to indicate the placement of the indicator with respect to the given node.
+enum Placement {
+  ///Placed above.
+  above,
+
+  ///Placed below.
+  below,
+
+  ///Placed as child.
+  child,
+}
 
 ///A widget used to render a treeview in the ui, based on the state of this tree inside the controller.
 class TreeView extends StatefulWidget {
@@ -38,7 +50,7 @@ class TreeView extends StatefulWidget {
     required this.nodeBuilder,
     this.indicator = const DefaultIndicator(height: 15),
     this.indicatorBuilder,
-    this.rowExtent = 8.0,
+    this.childExtent = 8.0,
     this.spacing = 8.0,
     this.animationDuration,
   });
@@ -58,8 +70,7 @@ class TreeView extends StatefulWidget {
   ///The builder provides 3 parameters:
   /// - BuildContext: The buildcontext this indicator is built in.
   /// - TreeNode: The node that the indicator uses as a reference to place itself.
-  ///   This can be either a sibling or a parent.
-  /// - Placement: the placement it will have in respect to [referenceNode].
+  /// - Placement: the placement it will have in respect to the given node.
   final Widget Function(
     BuildContext context,
     TreeNode referenceNode,
@@ -73,7 +84,7 @@ class TreeView extends StatefulWidget {
   final Widget indicator;
 
   ///The indent each child node has with respect to its parent.
-  final double rowExtent;
+  final double childExtent;
 
   ///The amount of space between nodes.
   final double spacing;
@@ -86,9 +97,9 @@ class TreeView extends StatefulWidget {
     itemBuilder: nodeBuilder,
     indicator: indicator,
     indicatorBuilder: indicatorBuilder,
-    rowExtent: rowExtent,
+    childExtent: childExtent,
     spacing: spacing,
-    animationDuration: animationDuration ?? Duration(milliseconds: 150),
+    animationDuration: animationDuration ?? const Duration(milliseconds: 150),
   );
 
   @override
@@ -107,7 +118,8 @@ class _TreeViewState extends State<TreeView> {
   @override
   void didUpdateWidget(covariant TreeView oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (oldWidget.controller != widget.controller) return;
+    if (oldWidget.controller == widget.controller) return;
+
     oldWidget.controller.removeListener(listener);
     widget.controller.addListener(listener);
   }
@@ -124,6 +136,7 @@ class _TreeViewState extends State<TreeView> {
       itemCount: widget.controller.rootCount,
       itemBuilder:
           (context, index) => _NodeWidget(
+            key: ValueKey(widget.controller.rootNodes[index].identifier),
             node: widget.controller.rootNodes[index],
             props: widget._props,
           ),
@@ -132,7 +145,7 @@ class _TreeViewState extends State<TreeView> {
 }
 
 class _NodeWidget extends StatefulWidget {
-  const _NodeWidget({required this.node, required this.props});
+  const _NodeWidget({super.key, required this.node, required this.props});
 
   final TreeNode node;
   final _TreeViewProps props;
@@ -154,16 +167,20 @@ class _NodeWidgetState extends State<_NodeWidget> {
     final offset = getGlobalOffset();
     final size = getObjectSize();
 
-    final isAbove = details.offset.dy < offset!.dy + size!.height / 3;
+    final isAbove = details.offset.dy < offset!.dy;
 
-    if (isAbove) {
+    final isChild = details.offset.dx > offset.dx + size!.width / 2;
+
+    if (isChild) {
+      return Placement.child;
+    } else if (isAbove) {
       return Placement.above;
     } else {
       return Placement.below;
     }
   }
 
-  void _resetDrag() {
+  void resetDrag() {
     widget.node.expand(notify: false);
     widget.node.isBeingDragged = false;
     isBeingDragged = false;
@@ -180,19 +197,33 @@ class _NodeWidgetState extends State<_NodeWidget> {
           DragTarget<TreeNode>(
             onAcceptWithDetails: (details) {
               final placement = getIndicatorPlacement(details);
-
-              if (placement == Placement.below) {
-                widget.props.controller.move(
-                  details.data,
-                  widget.node.index + 1,
-                  newParent: widget.node.parent,
-                );
+              if (placement == Placement.child) {
+                details.data.isAttached
+                    ? widget.props.controller.move(
+                      details.data,
+                      0,
+                      newParent: widget.node,
+                    )
+                    : widget.node.attachChild(details.data, index: 0);
+              } else if (placement == Placement.below) {
+                details.data.isAttached
+                    ? widget.props.controller.move(
+                      details.data,
+                      widget.node.index + 1,
+                      newParent: widget.node.parent,
+                    )
+                    : widget.node.attachSibling(details.data);
               } else {
-                widget.props.controller.move(
-                  details.data,
-                  widget.node.index,
-                  newParent: widget.node.parent,
-                );
+                details.data.isAttached
+                    ? widget.props.controller.move(
+                      details.data,
+                      widget.node.index,
+                      newParent: widget.node.parent,
+                    )
+                    : widget.node.attachSibling(
+                      details.data,
+                      index: widget.node.index,
+                    );
               }
             },
             onMove: (details) {
@@ -213,13 +244,13 @@ class _NodeWidgetState extends State<_NodeWidget> {
                     ? Draggable(
                       data: widget.node,
                       onDragStarted: () {
-                        widget.node.collapse();
                         widget.node.isBeingDragged = true;
                         isBeingDragged = true;
+                        widget.node.collapse();
                       },
-                      onDragEnd: (_) => _resetDrag(),
-                      onDraggableCanceled: (_, __) => _resetDrag(),
-                      onDragCompleted: () => _resetDrag(),
+                      onDragEnd: (_) => resetDrag(),
+                      onDraggableCanceled: (_, __) => resetDrag(),
+                      onDragCompleted: () => resetDrag(),
                       feedback: Material(
                         child: SizedBox(
                           width: getObjectSize()?.width ?? 500,
@@ -242,6 +273,17 @@ class _NodeWidgetState extends State<_NodeWidget> {
                         ) ??
                         widget.props.indicator,
                   widget.props.itemBuilder(context, widget.node),
+                  if (indicatorPlacement == Placement.child)
+                    Padding(
+                      padding: EdgeInsets.only(left: widget.props.childExtent),
+                      child:
+                          widget.props.indicatorBuilder?.call(
+                            context,
+                            widget.node,
+                            indicatorPlacement!,
+                          ) ??
+                          widget.props.indicator,
+                    ),
                   if (indicatorPlacement == Placement.below)
                     widget.props.indicatorBuilder?.call(
                           context,
@@ -263,12 +305,16 @@ class _NodeWidgetState extends State<_NodeWidget> {
             child:
                 widget.node.expanded && !isBeingDragged
                     ? Padding(
-                      padding: EdgeInsets.only(left: widget.props.rowExtent),
+                      padding: EdgeInsets.only(left: widget.props.childExtent),
                       child: Column(
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
                           for (final child in widget.node.children)
-                            _NodeWidget(node: child, props: widget.props),
+                            _NodeWidget(
+                              key: ValueKey(child.identifier),
+                              node: child,
+                              props: widget.props,
+                            ),
                         ],
                       ),
                     )